+++ CHANGELOG.md
@@ -30,6 +30,22 @@
     {% mytag 'John' msg='Hello' 123 %}
     ```
 
+- When a component tag receives multiple kwargs with the same name, it no longer raises `TypeError`.
+
+  Instead, the later kwargs overwrite the earlier ones.
+
+    ```django
+    {% mytag 'John' x=123 x=456 %}
+    ```
+
+- All template tags (`{% component %}`, `{% slot %}`, etc.) now include the exact tag (as found in the template) in the error message when an error occurs:
+
+    ```py
+    TypeError: Error in mytag: missing 1 required keyword-only arguments: 'msg'
+        1 | {% mytag 'John' %}
+            ^^^^^^^^^^^^^^^^^^
+    ```
+
 ## v0.144.0
 
 #### Feat